Clean up test code and refactor rest classes into better package structure
