+++ src/test/java/org/jabref/rest/jabmap/resources/RootResourceTest.java
@@ -1,4 +1,4 @@
-package org.jabref.rest.resources;
+package org.jabref.rest.jabmap.resources;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -10,14 +10,6 @@ import javax.ws.rs.client.Entity;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import com.microsoft.applicationinsights.core.dependencies.http.util.EntityUtils;
-import javafx.application.Platform;
-import javafx.collections.ObservableList;
-
-import org.jabref.gui.Globals;
-import org.jabref.gui.StateManager;
-import org.jabref.logic.importer.fileformat.endnote.Database;
-import org.jabref.model.database.BibDatabase;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.InternalField;
 import org.jabref.model.entry.field.MindMapField;
@@ -26,6 +18,7 @@ import org.jabref.model.entry.types.StandardEntryType;
 import org.jabref.model.jabmap.EdgeDirection;
 import org.jabref.model.jabmap.MindMapEdge;
 import org.jabref.model.jabmap.MindMapNode;
+import org.jabref.rest.jabmap.JabMapHTTPServer;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
@@ -37,7 +30,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class RootResourceTest {
 
-    private static String WEB_SERVICE_URI = "http://localhost:9898/";
+    private static final String WEB_SERVICE_URI = "http://localhost:9898/";
     private static Client client;
 
     @BeforeAll
@@ -67,6 +60,7 @@ class RootResourceTest {
             assertEquals(200, response.getStatus());
             String responseBody = response.readEntity(String.class);
             String expectedJson = "[{\"type\":{\"bibtex_metadata\":\"Article\",\"key\":\"article1\"}},{\"type\":{\"bibtex_metadata\":\"Book\",\"key\":\"book1\"}},{\"type\":{\"bibtex_metadata\":\"MastersThesis\",\"key\":\"MastersThesis1\"}}]";
+
             assertEquals(expectedJson, responseBody);
         } finally {
             response.close();
@@ -84,6 +78,7 @@ class RootResourceTest {
             assertEquals(200, response.getStatus());
             String responseBody = response.readEntity(String.class);
             String expectedJson = "[]";
+
             assertEquals(expectedJson, responseBody);
         } finally {
             response.close();
@@ -91,7 +86,7 @@ class RootResourceTest {
     }
 
     @Test
-    void getMindMapAndReferenceEntries() {
+    void getReferenceAndMindMapEntries() {
         List<BibEntry> bibEntries = setUpGetMindMapAndReferenceEntries();
         RootResource.databaseAccess.getActiveDatabase().insertEntries(bibEntries);
 
@@ -101,6 +96,7 @@ class RootResourceTest {
             assertEquals(200, response.getStatus());
             String responseBody = response.readEntity(String.class);
             String expectedJson = "[{\"type\":{\"bibtex_metadata\":\"Article\",\"key\":\"article1\"}},{\"type\":{\"bibtex_metadata\":\"Book\",\"key\":\"book1\"}},{\"type\":{\"bibtex_metadata\":\"MastersThesis\",\"key\":\"MastersThesis1\"}}]";
+
             assertEquals(expectedJson, responseBody);
         } finally {
             response.close();
@@ -118,6 +114,7 @@ class RootResourceTest {
             assertEquals(200, response.getStatus());
             String responseBody = response.readEntity(String.class);
             String expectedJson = "[]";
+
             assertEquals(expectedJson, responseBody);
         } finally {
             response.close();
@@ -125,43 +122,35 @@ class RootResourceTest {
     }
 
     @Test
-    void getRequestSingleNodeMindMap() {
+    void getSingleNodeMindMap() {
         BibEntry bibEntry = setupExpectedForSaveEmptyMap();
-        RootResource.databaseAccess.getActiveDatabase().insertEntries(bibEntry); // Add single node mindmap to the db
+        RootResource.databaseAccess.getActiveDatabase().insertEntries(bibEntry);
 
         Response response = null;
         try {
-            // Add single node map to the db
             response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request()
                     .put(Entity.json("{\"nodes\":[{\"id\":1,\"label\":\"New Map\",\"icons\":[],\"x_pos\":0,\"y_pos\":0}],\"edges\":[]}"));
 
-            // Make get request to retrieve map
-            response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request()
-                    .get();
+            response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request().get();
 
             String responseBody = response.readEntity(String.class);
             String expectedJson = "{\"nodes\":[{\"id\":1,\"label\":\"New Map\",\"icons\":[],\"x_pos\":0,\"y_pos\":0}],\"edges\":[]}";
 
             assertEquals(200, response.getStatus());
-
             assertEquals(expectedJson, responseBody);
-
         } finally {
-            // Close the Response object.
             response.close();
         }
     }
 
     @Test
-    void getRequestCompleteMindMap() {
+    void getCompleteMindMap() {
         List<BibEntry> bibEntries = setupExpectedForSaveCompleteMap();
-        RootResource.databaseAccess.getActiveDatabase().insertEntries(bibEntries); // Add complete mindmap to the db
+        RootResource.databaseAccess.getActiveDatabase().insertEntries(bibEntries);
 
         Response response = null;
         try {
-            // Make get request to retrieve map
-            response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request()
-                    .get();
+            response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request().get();
 
             String responseBody = response.readEntity(String.class);
             String expectedJson = "{\"nodes\":[{\"id\":1,\"label\":\"node 1\",\"icons\":[],\"x_pos\":0,\"y_pos\":0}," +
@@ -169,45 +158,36 @@ class RootResourceTest {
                     "\"edges\":[{\"node1_Id\":1,\"node2_Id\":2,\"label\":\"test edge\",\"direction\":\"DEFAULT\"}]}";
 
             assertEquals(200, response.getStatus());
-
             assertEquals(expectedJson, responseBody);
-
         } finally {
-            // Close the Response object.
             response.close();
         }
     }
 
     @Test
-    void getRequestBibAndMapEntries() {
-        List<BibEntry> bibEntries = setUpGetMindMapAndReferenceEntries();;
+    void getReferenceAndMapEntriesMindMap() {
+        List<BibEntry> bibEntries = setUpGetMindMapAndReferenceEntries();
         RootResource.databaseAccess.getActiveDatabase().insertEntries(bibEntries); // Add reference and mindmap entries to the db
 
         Response response = null;
         try {
-            // Make get request to retrieve map
-            response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request()
-                    .get();
+            response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request().get();
 
             String responseBody = response.readEntity(String.class);
             String expectedJson = "{\"nodes\":[{\"id\":3,\"label\":\"node 3\",\"icons\":[],\"x_pos\":10,\"y_pos\":10}]," +
                     "\"edges\":[{\"node1_Id\":1,\"node2_Id\":3,\"label\":\"test update edge\",\"direction\":\"LEFT\"}]}";
 
             assertEquals(200, response.getStatus());
-
             assertEquals(expectedJson, responseBody);
-
         } finally {
-            // Close the Response object.
             response.close();
         }
     }
 
     @Test
-    void getRequestNoMindMaps() {
+    void getEmptyMindMap() {
         Response response = null;
         try {
-            // Make get request to retrieve map
             response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request()
                     .get();
 
@@ -215,23 +195,19 @@ class RootResourceTest {
             String expectedJson = "{\"nodes\":[{\"label\":\"New Map\",\"icons\":[],\"x_pos\":0,\"y_pos\":0}],\"edges\":[]}";
 
             assertEquals(200, response.getStatus());
-
             assertEquals(expectedJson, responseBody);
-
         } finally {
-            // Close the Response object.
             response.close();
         }
     }
 
     @Test
-    void getRequestOnlyBibEntries() {
+    void getOnlyBibEntriesMindMap() {
         List<BibEntry> bibEntries = setUpEntriesForGetReferenceEntries();
         RootResource.databaseAccess.getActiveDatabase().insertEntries(bibEntries); // Add reference entries to the db
 
         Response response = null;
         try {
-            // Make get request to retrieve map
             response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request()
                     .get();
 
@@ -239,11 +215,8 @@ class RootResourceTest {
             String expectedJson = "{\"nodes\":[{\"label\":\"New Map\",\"icons\":[],\"x_pos\":0,\"y_pos\":0}],\"edges\":[]}";
 
             assertEquals(200, response.getStatus());
-
             assertEquals(expectedJson, responseBody);
-
         } finally {
-            // Close the Response object.
             response.close();
         }
     }
@@ -254,8 +227,6 @@ class RootResourceTest {
 
         Response response = null;
         try {
-            // Make an invocation on a Concert URI and specify Java-
-            // serialization as the required data format.
             response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request()
                              .put(Entity.json("{\n" +
                                      "  \"nodes\": [\n" +
@@ -285,14 +256,12 @@ class RootResourceTest {
             assertEquals(200, response.getStatus());
 
             ArrayList<BibEntry> actualBibEntries = new ArrayList<>(RootResource.databaseAccess.getActiveDatabase().getEntries());
-
             assertEquals(expectedBibEntries.size(), actualBibEntries.size());
             for (BibEntry bibEntry : expectedBibEntries) {
                 assertTrue(actualBibEntries.contains(bibEntry));
                 actualBibEntries.remove(bibEntry);
             }
         } finally {
-            // Close the Response object.
             response.close();
         }
     }
@@ -303,8 +272,6 @@ class RootResourceTest {
 
         Response response = null;
         try {
-            // Make an invocation on a Concert URI and specify Java-
-            // serialization as the required data format.
             response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request()
                              .put(Entity.json("{\n" +
                                      "  \"nodes\": [\n" +
@@ -321,11 +288,9 @@ class RootResourceTest {
             assertEquals(200, response.getStatus());
 
             List<BibEntry> actualBibEntries = RootResource.databaseAccess.getActiveDatabase().getEntries();
-
             assertEquals(1, actualBibEntries.size());
             assertEquals(expectedBibEntry, actualBibEntries.get(0));
         } finally {
-            // Close the Response object.
             response.close();
         }
     }
@@ -336,8 +301,6 @@ class RootResourceTest {
 
         Response response = null;
         try {
-            // Make an invocation on a Concert URI and specify Java-
-            // serialization as the required data format.
             response = client.target(WEB_SERVICE_URI).path("libraries/current/map").request()
                              .put(Entity.json("{\n" +
                                      "  \"nodes\": [\n" +
@@ -405,15 +368,12 @@ class RootResourceTest {
             assertEquals(200, response.getStatus());
 
             ArrayList<BibEntry> actualBibEntries = new ArrayList<>(RootResource.databaseAccess.getActiveDatabase().getEntries());
-
             assertEquals(expectedBibEntries.size(), actualBibEntries.size());
-
             for (BibEntry bibEntry : expectedBibEntries) {
                 assertTrue(actualBibEntries.contains(bibEntry));
                 actualBibEntries.remove(bibEntry);
             }
         } finally {
-            // Close the Response object.
             response.close();
         }
     }
@@ -421,21 +381,18 @@ class RootResourceTest {
     private List<BibEntry> setupExpectedForSaveOverrideMap() {
         BibEntry node1 = new BibEntry(MindMapEntryType.Node)
                 .withField(MindMapField.NODE_LABEL, "my updated node")
-                // .withField(MindMapField.NODE_ICONS,iconstr)
                 .withField(InternalField.KEY_FIELD, MindMapNode.getCitationKeyFromId(1L))
                 .withField(MindMapField.NODE_XPOS, String.valueOf(40))
                 .withField(MindMapField.NODE_YPOS, String.valueOf(20));
 
         BibEntry node2 = new BibEntry(MindMapEntryType.Node)
                 .withField(MindMapField.NODE_LABEL, "node 2")
-                // .withField(MindMapField.NODE_ICONS,iconstr)
                 .withField(InternalField.KEY_FIELD, MindMapNode.getCitationKeyFromId(2L))
                 .withField(MindMapField.NODE_XPOS, String.valueOf(10))
                 .withField(MindMapField.NODE_YPOS, String.valueOf(0));
 
         BibEntry node3 = new BibEntry(MindMapEntryType.Node)
                 .withField(MindMapField.NODE_LABEL, "node 3")
-                // .withField(MindMapField.NODE_ICONS,iconstr)
                 .withField(InternalField.KEY_FIELD, MindMapNode.getCitationKeyFromId(3L))
                 .withField(MindMapField.NODE_XPOS, String.valueOf(10))
                 .withField(MindMapField.NODE_YPOS, String.valueOf(10));
@@ -456,7 +413,6 @@ class RootResourceTest {
     private BibEntry setupExpectedForSaveEmptyMap() {
         BibEntry node1 = new BibEntry(MindMapEntryType.Node)
                 .withField(MindMapField.NODE_LABEL, "New Map")
-                // .withField(MindMapField.NODE_ICONS,iconstr)
                 .withField(InternalField.KEY_FIELD, MindMapNode.getCitationKeyFromId(1L))
                 .withField(MindMapField.NODE_XPOS, String.valueOf(0))
                 .withField(MindMapField.NODE_YPOS, String.valueOf(0));
@@ -466,14 +422,12 @@ class RootResourceTest {
     private List<BibEntry> setupExpectedForSaveCompleteMap() {
         BibEntry node1 = new BibEntry(MindMapEntryType.Node)
                 .withField(MindMapField.NODE_LABEL, "node 1")
-                // .withField(MindMapField.NODE_ICONS,iconstr)
                 .withField(InternalField.KEY_FIELD, MindMapNode.getCitationKeyFromId(1L))
                 .withField(MindMapField.NODE_XPOS, String.valueOf(0))
                 .withField(MindMapField.NODE_YPOS, String.valueOf(0));
 
         BibEntry node2 = new BibEntry(MindMapEntryType.Node)
                 .withField(MindMapField.NODE_LABEL, "node 2")
-                // .withField(MindMapField.NODE_ICONS,iconstr)
                 .withField(InternalField.KEY_FIELD, MindMapNode.getCitationKeyFromId(2L))
                 .withField(MindMapField.NODE_XPOS, String.valueOf(10))
                 .withField(MindMapField.NODE_YPOS, String.valueOf(0));
@@ -511,7 +465,6 @@ class RootResourceTest {
 
         BibEntry node3 = new BibEntry(MindMapEntryType.Node)
                 .withField(MindMapField.NODE_LABEL, "node 3")
-                // .withField(MindMapField.NODE_ICONS,iconstr)
                 .withField(InternalField.KEY_FIELD, MindMapNode.getCitationKeyFromId(3L))
                 .withField(MindMapField.NODE_XPOS, String.valueOf(10))
                 .withField(MindMapField.NODE_YPOS, String.valueOf(10));
